Migrate LogEntry.object_id to bigint (#5520)

+++ src/pretix/base/models/log.py
@@ -66,7 +66,7 @@ class LogEntry(models.Model):
     :type data: str
     """
     content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
-    object_id = models.PositiveIntegerField(db_index=True)
+    object_id = models.PositiveBigIntegerField(db_index=True)
     content_object = GenericForeignKey('content_type', 'object_id')
     datetime = models.DateTimeField(auto_now_add=True)
     user = models.ForeignKey('User', null=True, blank=True, on_delete=models.PROTECT)
